+++ example/lib/markers/pages/hide-markers.page.dart
@@ -9,7 +9,8 @@ import 'package:visual_editor/visual-editor.dart';
 import '../../shared/const/dimensions.const.dart';
 import '../../shared/widgets/demo-page-scaffold.dart';
 import '../../shared/widgets/loading.dart';
-import '../models/markers-and-scroll-offset.model.dart';
+import '../models/marker-and-relative-position.model.dart';
+import '../models/markers-attachments-position.dart';
 import '../widgets/markers-attachments.dart';
 import 'delete-markers.page.dart';
 
@@ -32,7 +33,7 @@ class _HideMarkersPageState extends State<HideMarkersPage> {
   // We will update only the MarkersAttachmentsSidebar via the stream.
   // By using this trick we can prevent Flutter from running expensive page updates.
   // We will target our updates only on the area that renders the attachments (far better performance).
-  final _markers$ = StreamController<MarkersAndScrollOffset>.broadcast();
+  final _markers$ = StreamController<MarkersAttachmentsPos>.broadcast();
 
   @override
   void initState() {
@@ -198,9 +199,19 @@ class _HideMarkersPageState extends State<HideMarkersPage> {
   // (!) Avoid setState() on the parent page, setState in a smallest possible widget to minimise the update cost.
   void _updateMarkerAttachments() {
     final markers = _controller?.getAllMarkers() ?? [];
+    final markersAndPos = <MarkerAndRelPos>[];
+
+    markers.forEach((marker) {
+      markersAndPos.add(
+        MarkerAndRelPos(
+          marker: marker,
+        ),
+      );
+    });
+
     _markers$.sink.add(
-      MarkersAndScrollOffset(
-        markers: markers,
+      MarkersAttachmentsPos(
+        markers: markersAndPos,
         scrollOffset: SCROLLABLE ? _scrollController.offset : 0,
       ),
     );